srcs/builtinのリファクタリングとノーム対応
