+++ srcs/builtin/builtin_echo.c
@@ -1,53 +1,38 @@
 #include "minishell.h"
 
+void	put_argv(char **argv, int i)
+{
+	while (argv[i] != NULL)
+	{
+		ms_putstr_fd(argv[i], STDOUT);
+		i++;
+		if (argv[i] != NULL)
+			ms_putstr_fd(" ", STDOUT);
+	}
+}
+
 int	builtin_echo(int argc, char *argv[])
 {
-	int		flag;
-	size_t	i;
+	t_boolean		is_flag;
+	size_t			i;
 
 	if (argc == 1)
-	{
 		ms_putchar_fd('\n', STDOUT);
-	}
 	else
 	{
 		if (ms_strcmp(argv[1], "-n") == 0)
 		{
-			flag = 1;
+			is_flag = TRUE;
 			i = 2;
 		}
 		else
 		{
-			flag = 0;
+			is_flag = FALSE;
 			i = 1;
 		}
-		while (argv[i] != NULL)
-		{
-			ms_putstr_fd(argv[i], STDOUT);
-			i++;
-			if (argv[i] != NULL)
-				ms_putstr_fd(" ", STDOUT);
-		}
-		if (flag == 0)
+		put_argv(argv, i);
+		if (is_flag == FALSE)
 			ms_putstr_fd("\n", STDOUT);
 	}
-	return (0);
+	return (EXIT_SUCCESS);
 }
-
-/*int main(void)
-{
-	char	*test_1[] = {"cat", NULL};
-	char	*test_2[] = {"cat", "taiyou", NULL};
-	char	*test_3[] = {"cat", "taiyou", "sekiguchi", NULL};
-	char	*test_4[] = {"cat", "-n", NULL};
-	char	*test_5[] = {"cat", "-n", "taiyou", NULL};
-	char	*test_6[] = {"cat", "-n", "taiyou", "sekiguchi", NULL};
-
-	echo(1, test_1);
-	echo(2, test_2);
-	echo(3, test_3);
-	echo(2, test_4);
-	echo(3, test_5);
-	echo(4, test_6);
-	return (0);
-}*/
